fixed build bugs
