+++ .gitignore
@@ -7,7 +7,6 @@ node_modules
 package-lock.json
 # testing
 coverage
-
 # next.js
 .next/
 out/
@@ -35,3 +34,7 @@ yarn-error.log*
 # vercel
 .vercel
 *wrangler.toml
+
+
+target
+pkg
