+++ core/Cargo.toml
@@ -39,7 +39,7 @@ lto = true
 [dev-dependencies]
 criterion = {version = "0.4.0", features= ["html_reports"]}
 reqwest = "0.11.16"
-[[bench]]
-name = "benchmarks"
-harness = false
+# [[bench]]
+# name = "benchmarks"
+# harness = false
