+++ package.json
@@ -14,7 +14,7 @@
   },
   "packageManager": "npm@9.6.4",
   "workspaces": [
-    "admin",
+    "dash",
     "api",
     "core"
   ]
